Refactor engine.rs - Split 711-line file into logical modules
## Summary

Refactored the monolithic engine.rs (711 lines) into a well-organized module structure following the vertical slice approach. Each module now represents a distinct responsibility in the game loop, improving maintainability and code organization.

## Changes

### Module Structure
- Created `bucket-brigade-core/src/engine/` directory with 7 files:
  - `mod.rs` (~12 lines): Module declarations and public API exports
  - `types.rs` (~32 lines): Core type definitions (GameState, GameResult, StepResult)
  - `core.rs` (~120 lines): BucketBrigade struct, new(), reset(), step(), check_termination()
  - `phases.rs` (~65 lines): Fire mechanics (extinguish, spread, burn_out, spontaneous_ignition)
  - `rewards.rs` (~53 lines): Reward computation (per-step and final rewards)
  - `observation.rs` (~75 lines): Trajectory recording and state observation methods
  - `tests.rs` (~385 lines): All 28 test functions (organized by functionality)

### Benefits
- **Improved Maintainability**: Each module under 200 lines, clear separation of concerns
- **Better Organization**: Fire mechanics, rewards, and observations isolated in dedicated modules
- **Enhanced Readability**: Easier to navigate and understand specific subsystems
- **Module Visibility**: Used `pub(super)` for internal APIs, maintaining encapsulation
- **No Behavioral Changes**: All 59 tests pass, cargo clippy clean, public API unchanged

### Testing
- ✅ All 28 engine tests pass
- ✅ All 59 total tests pass (engine + rng + scenarios + types)
- ✅ Cargo clippy clean (no warnings)
- ✅ Cargo fmt applied
- ✅ No public API changes (backward compatible)

Closes #79
