+++ bucket-brigade-core/src/scenarios.rs
@@ -3,24 +3,24 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Scenario {
     // Fire dynamics
-    pub prob_fire_spreads_to_neighbor: f32,  // Probability fire spreads to adjacent house
-    pub prob_solo_agent_extinguishes_fire: f32,  // Probability one agent extinguishes fire
-    pub prob_house_catches_fire: f32,  // Probability house catches fire each night
+    pub prob_fire_spreads_to_neighbor: f32, // Probability fire spreads to adjacent house
+    pub prob_solo_agent_extinguishes_fire: f32, // Probability one agent extinguishes fire
+    pub prob_house_catches_fire: f32,       // Probability house catches fire each night
 
     // Team scoring (collective outcome)
-    pub team_reward_house_survives: f32,  // Team reward for each house that survives
-    pub team_penalty_house_burns: f32,    // Team penalty for each house that burns
+    pub team_reward_house_survives: f32, // Team reward for each house that survives
+    pub team_penalty_house_burns: f32,   // Team penalty for each house that burns
 
     // Individual rewards (ownership-based, for future use in issue #52)
-    pub reward_own_house_survives: f32,      // Individual reward when own house survives
-    pub reward_other_house_survives: f32,    // Individual reward when other house survives
-    pub penalty_own_house_burns: f32,        // Individual penalty when own house burns
-    pub penalty_other_house_burns: f32,      // Individual penalty when other house burns
+    pub reward_own_house_survives: f32, // Individual reward when own house survives
+    pub reward_other_house_survives: f32, // Individual reward when other house survives
+    pub penalty_own_house_burns: f32,   // Individual penalty when own house burns
+    pub penalty_other_house_burns: f32, // Individual penalty when other house burns
 
     // Costs and structure
-    pub cost_to_work_one_night: f32,  // Cost incurred when agent chooses to work
-    pub min_nights: u32,              // Minimum nights before game can end
-    pub num_agents: usize,            // Number of agents in game
+    pub cost_to_work_one_night: f32, // Cost incurred when agent chooses to work
+    pub min_nights: u32,             // Minimum nights before game can end
+    pub num_agents: usize,           // Number of agents in game
 }
 
 pub const SCENARIOS: phf::Map<&'static str, Scenario> = phf::phf_map! {
@@ -142,7 +142,11 @@ mod tests {
     #[test]
     fn test_all_scenarios_have_4_agents() {
         for (name, scenario) in SCENARIOS.entries() {
-            assert_eq!(scenario.num_agents, 4, "Scenario '{}' should have 4 agents", name);
+            assert_eq!(
+                scenario.num_agents, 4,
+                "Scenario '{}' should have 4 agents",
+                name
+            );
         }
     }
 
@@ -150,16 +154,24 @@ mod tests {
     fn test_all_scenarios_valid_probabilities() {
         for (name, scenario) in SCENARIOS.entries() {
             assert!(
-                scenario.prob_fire_spreads_to_neighbor >= 0.0 && scenario.prob_fire_spreads_to_neighbor <= 1.0,
-                "Scenario '{}' has invalid prob_fire_spreads_to_neighbor: {}", name, scenario.prob_fire_spreads_to_neighbor
+                scenario.prob_fire_spreads_to_neighbor >= 0.0
+                    && scenario.prob_fire_spreads_to_neighbor <= 1.0,
+                "Scenario '{}' has invalid prob_fire_spreads_to_neighbor: {}",
+                name,
+                scenario.prob_fire_spreads_to_neighbor
             );
             assert!(
-                scenario.prob_solo_agent_extinguishes_fire >= 0.0 && scenario.prob_solo_agent_extinguishes_fire <= 1.0,
-                "Scenario '{}' has invalid prob_solo_agent_extinguishes_fire: {}", name, scenario.prob_solo_agent_extinguishes_fire
+                scenario.prob_solo_agent_extinguishes_fire >= 0.0
+                    && scenario.prob_solo_agent_extinguishes_fire <= 1.0,
+                "Scenario '{}' has invalid prob_solo_agent_extinguishes_fire: {}",
+                name,
+                scenario.prob_solo_agent_extinguishes_fire
             );
             assert!(
                 scenario.prob_house_catches_fire >= 0.0 && scenario.prob_house_catches_fire <= 1.0,
-                "Scenario '{}' has invalid prob_house_catches_fire: {}", name, scenario.prob_house_catches_fire
+                "Scenario '{}' has invalid prob_house_catches_fire: {}",
+                name,
+                scenario.prob_house_catches_fire
             );
         }
     }
@@ -179,7 +191,10 @@ mod tests {
         assert!(json.contains("\"num_agents\":4"));
 
         let deserialized: Scenario = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.prob_fire_spreads_to_neighbor, scenario.prob_fire_spreads_to_neighbor);
+        assert_eq!(
+            deserialized.prob_fire_spreads_to_neighbor,
+            scenario.prob_fire_spreads_to_neighbor
+        );
         assert_eq!(deserialized.num_agents, scenario.num_agents);
     }
 
@@ -194,11 +209,13 @@ mod tests {
         for (name, scenario) in SCENARIOS.entries() {
             assert!(
                 scenario.team_reward_house_survives > 0.0,
-                "Scenario '{}' should have positive reward for saved houses", name
+                "Scenario '{}' should have positive reward for saved houses",
+                name
             );
             assert!(
                 scenario.team_penalty_house_burns > 0.0,
-                "Scenario '{}' should have positive penalty for ruined houses", name
+                "Scenario '{}' should have positive penalty for ruined houses",
+                name
             );
         }
     }
